refactor(ds): abstract out logics from CustomFilter (#212)
* refactor(ds): abstract out logics

* fix(ds): add test

* refactor(ds): fix type error

* fix(ds): add test

* refactor(ds): refactor test

* refactor(ds): move test case

* refactor(ds): fix typo
